Restructure streams domain: narrative Stream DO, isolated connections, split contracts
- stream-durable-object.ts: narrative order (append commit point → core
  processor validateAppend/reduce/processEvent → checkpoint → subscriptions),
  storage and connection machinery extracted
- stream-storage.ts: chunked SQLite event log
- stream-connections.ts: connection table, pump, RPC retention (absorbs
  rpc-lifecycle.ts), idle teardown
- processor-contracts.ts: contract type machinery split from the
  StreamProcessor class; single defineProcessorContract signature; dead
  contract.buildEvent/class buildEvent surface removed
- stream-processor-host.ts: add() derives name from contract.slug;
  unified supersedeConnection path; wake handshake documented as the
  live-capability rhyme with domains/itx/live-capability.ts
- core-processor-contract.ts: shared payload schemas (no more duplicated
  subscription/rule payload definitions)
- domain README.md; stale withStreamProcessor CONTEXT.md entry fixed
