+++ apps/os/src/README.md
@@ -193,12 +193,15 @@ ends the loop — the completion event then carries no `result` key.
 `StreamDurableObject` owns the journal (DO SQLite); its storage methods stay
 synchronous internally while the public `Stream` capability is async through
 an RpcTarget. Processors are hosted by their domain DO via
-`createStreamProcessorHost(...)` and receive a full public `Stream` capability
-— never raw DO stubs. Subscription handshakes are identity-only: the stream
-tells the host which `subscriptionKey` to open; the host subscribes on its own
-stream capability. State is a fold of the journal; the `{offset, state}`
-checkpoint is a disposable cache (the doctrine:
-`docs/domain-objects-and-stream-processors.md`).
+`createStreamProcessorHost(...)` — `host.add((deps) => new SomeProcessor(deps))`
+— and receive a full public `Stream` capability, never raw DO stubs.
+Subscription handshakes are identity-only: the stream tells the host which
+`subscriptionKey` to open; the host subscribes on its own stream capability,
+handing the stream a live `processEventBatch` callback (the same
+live-capability shape as itx provision). State is a fold of the journal; the
+`{offset, state}` checkpoint is a disposable cache. The domain's own guide is
+`domains/streams/README.md`; the doctrine is
+`docs/domain-objects-and-stream-processors.md`.
 
 The browser stream mirror is a second host of the same engine: the dashboard
 keeps a local event table plus derived tables and runs real `StreamProcessor`
