+++ apps/os/src/domains/agents/agent-durable-object.ts
@@ -7,14 +7,10 @@ import {
 } from "../streams/stream-processor-host.ts";
 import { StreamProcessorRpcTarget } from "../../rpc-targets.ts";
 import { StreamRpcTarget } from "../../rpc-targets.ts";
-import { SlackAgentProcessorContract } from "../integrations/slack-agent-processor-contract.ts";
 import { SlackAgentProcessor } from "../integrations/slack-agent-processor-implementation.ts";
 import { callProjectSlackWebApi } from "../integrations/slack-api.ts";
-import { AgentProcessorContract } from "./agent-processor-contract.ts";
 import { AgentProcessor } from "./agent-processor-implementation.ts";
-import { CloudflareAiProcessorContract } from "./cloudflare-ai-processor-contract.ts";
 import { CloudflareAiProcessor } from "./cloudflare-ai-processor-implementation.ts";
-import { OpenAiWsProcessorContract } from "./openai-ws-processor-contract.ts";
 import { OpenAiWsProcessor } from "./openai-ws-processor-implementation.ts";
 import { parseAgentDurableObjectName, readOpenAiApiKeyFromAppConfig } from "./utils.ts";
 
@@ -28,12 +24,8 @@ export class AgentDurableObject extends DurableObject<Env> {
   readonly #processorHost = createStreamProcessorHost(this.ctx, {
     stream: this.#stream,
   });
-  readonly #agentProcessor = this.#processorHost.add(
-    AgentProcessorContract.slug,
-    (deps) => new AgentProcessor(deps),
-  );
+  readonly #agentProcessor = this.#processorHost.add((deps) => new AgentProcessor(deps));
   readonly cloudflareAiProcessor = this.#processorHost.add(
-    CloudflareAiProcessorContract.slug,
     (deps) =>
       new CloudflareAiProcessor({
         ...deps,
@@ -44,7 +36,6 @@ export class AgentDurableObject extends DurableObject<Env> {
   // Registered even without an OpenAI key: the processor then fails requests
   // with a clear llm-request-completed error instead of crashing the host.
   readonly openAiWsProcessor = this.#processorHost.add(
-    OpenAiWsProcessorContract.slug,
     (deps) =>
       new OpenAiWsProcessor({
         ...deps,
@@ -58,7 +49,6 @@ export class AgentDurableObject extends DurableObject<Env> {
   // subscription. Slack-facing side effects are best effort: a failed status
   // update or reaction must not wedge the processor checkpoint.
   readonly slackAgentProcessor = this.#processorHost.add(
-    SlackAgentProcessorContract.slug,
     (deps) =>
       new SlackAgentProcessor({
         ...deps,
