+++ apps/os/src/domains/agents/openai-ws-processor-contract.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { defineProcessorContract } from "../streams/stream-processor.ts";
+import { defineProcessorContract } from "../streams/processor-contracts.ts";
 import { AgentProcessorContract } from "./agent-processor-contract.ts";
 
 /** Default model for the openai-ws provider (matches the legacy os default). */
