+++ apps/os/src/domains/projects/project-durable-object.ts
@@ -16,11 +16,9 @@ import {
 import { deepRetainRpcStubs } from "../itx/live-capability.ts";
 import { readOpenAiApiKeyFromAppConfig } from "../agents/utils.ts";
 import { secretErrorResponse, secretReferencePathsFromHeaders } from "../secrets/utils.ts";
-import { SlackProcessorContract } from "../integrations/slack-processor-contract.ts";
 import { SlackProcessor } from "../integrations/slack-processor-implementation.ts";
 import { eyesReactionTargetFromWebhookPayload } from "../integrations/slack-agent-processor-implementation.ts";
 import { callProjectSlackWebApi } from "../integrations/slack-api.ts";
-import { ProjectProcessorContract } from "./project-processor-contract.ts";
 import { ProjectProcessor } from "./project-processor-implementation.ts";
 
 export class ProjectDurableObject extends DurableObject<Env> {
@@ -34,7 +32,6 @@ export class ProjectDurableObject extends DurableObject<Env> {
     }),
   });
   readonly #projectProcessor = this.#processorHost.add(
-    ProjectProcessorContract.slug,
     (deps) =>
       new ProjectProcessor({
         ...deps,
@@ -54,7 +51,7 @@ export class ProjectDurableObject extends DurableObject<Env> {
   // instance addressed at `/integrations/slack` (the host stream is this DO's
   // own path stream), where the OAuth connect / project bootstrap configured
   // its subscription; registering it on every instance is harmless.
-  readonly #slackProcessor = this.#processorHost.add(SlackProcessorContract.slug, (deps) => {
+  readonly #slackProcessor = this.#processorHost.add((deps) => {
     return new SlackProcessor({
       ...deps,
       acknowledgeRoutedWebhook: async ({ payload }) => {
