+++ apps/os/src/domains/secrets/secret-processor-contract.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { defineProcessorContract } from "../streams/stream-processor.ts";
+import { defineProcessorContract } from "../streams/processor-contracts.ts";
 
 export const SecretProcessorContract = defineProcessorContract({
   slug: "secret",
